+++ gcc/tree-ssa-structalias.cc
@@ -5885,7 +5885,8 @@ push_fields_onto_fieldstack (tree type, vec<fieldoff_s> *fieldstack,
 		&& offset + foff != 0)
 	      {
 		fieldoff_s e
-		  = {0, offset + foff, false, false, true, false, NULL_TREE};
+		  = {0, static_cast<unsigned long long>(offset + foff), false,
+		     false, true, false, NULL_TREE};
 		pair = fieldstack->safe_push (e);
 	      }
