+++ gcc/tree-vect-slp.cc
@@ -4615,7 +4615,8 @@ vect_prologue_cost_for_slp (slp_tree node,
       nelt_limit = const_nunits;
       hash_set<vect_scalar_ops_slice_hash> vector_ops;
       for (unsigned int i = 0; i < SLP_TREE_NUMBER_OF_VEC_STMTS (node); ++i)
-	if (!vector_ops.add ({ ops, i * const_nunits, const_nunits }))
+	if (!vector_ops.add ({ ops, static_cast<unsigned int>(i * const_nunits),
+	                       static_cast<unsigned int>(const_nunits) }))
 	  starts.quick_push (i * const_nunits);
     }
   else
